Logica APIs

+++ Aplicacion/app/src/main/java/com/madicine/deliverycontrol/MenuPrincipal.kt
@@ -1,7 +1,6 @@
 package com.madicine.deliverycontrol
 
 import android.content.Intent
-import android.content.SharedPreferences
 import android.os.Build
 import android.os.Bundle
 import android.widget.Button
@@ -10,16 +9,18 @@ import android.widget.TextView
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
+import androidx.activity.viewModels
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import androidx.lifecycle.Observer
 import com.google.firebase.auth.FirebaseAuth
 import com.madicine.deliverycontrol.Entities.Usuario
 import com.madicine.deliverycontrol.Interfaces.PermissionHandler
 import com.madicine.deliverycontrol.Utilities.PermissionsUtils
+import com.madicine.deliverycontrol.viewModels.MedicamentosViewModel
 import java.text.SimpleDateFormat
-import java.time.ZoneOffset
-import java.time.ZonedDateTime
 import java.util.Locale
 
 class MenuPrincipal : AppCompatActivity() {
@@ -27,11 +28,15 @@ class MenuPrincipal : AppCompatActivity() {
     private lateinit var permissionHandler: PermissionHandler
     private lateinit var tl_usuario: TextView
     private lateinit var tl_fecha: TextView
-    private lateinit var tlCodigo: TextView
+    private lateinit var tlDatos: TextView
     private lateinit var btnConsulta: Button
     private lateinit var btnILogOut: ImageButton
     private lateinit var barcodeScannerLauncher: ActivityResultLauncher<Intent>
     private var usuario: Usuario? = null
+    private var scannedCode : String? = null
+
+    //Declaracion de variable ViewModel
+    private val viewModel: MedicamentosViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -53,6 +58,7 @@ class MenuPrincipal : AppCompatActivity() {
 
         tl_usuario = findViewById(R.id.tlUsuario);
         tl_fecha = findViewById(R.id.tlFecha)
+        tlDatos = findViewById(R.id.tlDatos)
         btnILogOut = findViewById(R.id.imgLogOut)
         btnConsulta = findViewById(R.id.btnConsulta)
 
@@ -67,7 +73,6 @@ class MenuPrincipal : AppCompatActivity() {
         val nombre = usuario?.nombre + " " + usuario?.apellido
         val email = usuario?.email
         val uid = usuario?.udi
-        val provider = usuario?.udi
         val sdf = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault())
         val currentDate = sdf.format(System.currentTimeMillis())
 
@@ -85,17 +90,29 @@ class MenuPrincipal : AppCompatActivity() {
         tl_usuario.text = nombre
         tl_fecha.text = currentDate
 
-        //Configuración
-        setup()
-
+        //Se obtienen los resultados del scanner
         barcodeScannerLauncher = registerForActivityResult(
             ActivityResultContracts.StartActivityForResult()
         ) { result ->
             if (result.resultCode == RESULT_OK) {
-                val scannedCode = result.data?.getStringExtra("scanned_code")
-                tlCodigo.text = scannedCode ?: "No se pudo escanear el código"
+                scannedCode = result.data?.getStringExtra("scanned_code")
+                scannedCode?.let {
+                    viewModel.enviarCodigoBarras(it,uid.toString())
+                }
             }
         }
+
+        //Lectura de ViewModel
+        viewModel.respuesta.observe(this, Observer { respuesta ->
+            respuesta?.let {
+                tlDatos.text = it
+            } ?: run {
+                tlDatos.text = "Error al obtener la respuesta"
+            }
+        })
+
+        //Configuración
+        setup()
     }
 
     private fun setup() {
@@ -108,6 +125,20 @@ class MenuPrincipal : AppCompatActivity() {
         }
 
         btnILogOut.setOnClickListener {
+            logOut()
+        }
+    }
+
+    /**
+     * Metodo para cerrar sesión
+     */
+    private fun logOut(){
+        val builder = AlertDialog.Builder(this)
+
+        builder.setMessage(R.string.alertLogOut)
+
+        // Configuración del botón "Sí"
+        builder.setPositiveButton("Sí") { dialog, _ ->
             //Borrado de datos de SharedPreferenses
             val prefs = getSharedPreferences(getString(R.string.prefs_file), MODE_PRIVATE).edit()
             prefs.clear()
@@ -116,6 +147,16 @@ class MenuPrincipal : AppCompatActivity() {
             //Cerrar sesión de firebase
             FirebaseAuth.getInstance().signOut()
             onBackPressed()
+
+            dialog.dismiss()
         }
+
+        // Configuración del botón "No"
+        builder.setNegativeButton("No") { dialog, _ ->
+            dialog.dismiss()
+        }
+
+        val dialog: AlertDialog = builder.create()
+        dialog.show()
     }
 }
